调整新爬虫程序为自动加载，之前是需要配置文件配置方法名。不方便。 同时，降暂时失效和调试中的爬虫，移到configFetcherBak里面，避免启动时候自动加载。
